:art: Move Topic class to the same script as TopicLogger.
